% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbams.R
\name{dap.sbams}
\alias{dap.sbams}
\title{Integrative Genetic Association Analysis using Deterministic Approximation of Posteriors}
\usage{
dap.sbams(sbams)
}
\arguments{
\item{sbams}{a sbams object}
}
\value{
a dap object including model summary, SNP summary and signal cluster summary
}
\description{
The \code{dap.sbams} function is a simpler wrapper for \code{dap} function on sbams objects.
}
