% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbams.R
\name{read.sbams}
\alias{read.sbams}
\title{Read a sbams file}
\usage{
read.sbams(file, normalize = TRUE)
}
\arguments{
\item{file}{file path to the sbams file}

\item{normalize}{TRUE by default. Data should be normalized before analysis.}
}
\value{
a sbams object
}
\description{
Read a sbams file
}
