% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.dap}
\alias{summary.dap}
\title{Summary the dap analysis result}
\usage{
\method{summary}{dap}(object, ...)
}
\arguments{
\item{object}{a dap object}
}
\description{
Summary the dap analysis result
}
