% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbams.R
\name{summary.sbams}
\alias{summary.sbams}
\title{summary sbams object}
\usage{
\method{summary}{sbams}(sbams)
}
\arguments{
\item{sbams}{a sbams object}
}
\value{
It will show the sample size [n] and total sum of squares for the quantitative trait [syy],
and write files [LD.dat] and [est.dat] under the working directory.
}
\description{
\code{summary.sbams} helps to generate the summary statistics from a sbams object
to facilitate downstream analysis.
}
